% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5mincent_surv.R
\name{base_5mincent_surv}
\alias{base_5mincent_surv}
\title{base_5mincent_surv}
\format{
a ggplot object with the strata as a geom_sf() object
}
\description{
Lake Erie 5-minute grid centroids within survey area as base ggplot object
}
\examples{
\dontrun{
data("base_5mincent_surv")
base_5mincent_surv
}

}
\seealso{
shape_5mincent_surv
}
