% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5mingrid.R
\name{base_5mingrid}
\alias{base_5mingrid}
\title{base_5mingrid}
\format{
a ggplot object with the strata as a geom_sf() object
}
\description{
Lake Erie 5-minute grid as base ggplot object
}
\examples{
\dontrun{
data("base_5mingrid")
base_5mingrid
}

}
\seealso{
shape_5mingrid
}
