% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5mingrid.R
\name{base_5mingrid_surv}
\alias{base_5mingrid_surv}
\title{base_5mingrid_surv}
\format{
a ggplot object with the strata as a geom_sf() object
}
\description{
Lake Erie 5-minute grid within survey area as base ggplot object
}
\examples{
\dontrun{
data("base_5mingrid_surv")
base_5mingrid_surv
}

}
\seealso{
shape_5mingrid_surv
}
