% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbstrata.R
\name{base_cbstrata}
\alias{base_cbstrata}
\title{base_cbstrata}
\format{
a ggplot object with the strata as a geom_sf() object
}
\description{
Central Basin Strata as base ggplot object
}
\examples{
\dontrun{
data("base_cbstrata")
base_cbstrata
}

}
\seealso{
shape_cbstrata
}
