% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebstrata.R
\name{base_ebstrata}
\alias{base_ebstrata}
\title{base_ebstrata}
\format{
a ggplot object with the strata as a geom_sf() object
}
\description{
East Basin Strata as base ggplot object
}
\examples{
\dontrun{
data("base_ebstrata")
base_ebstrata
}

}
\seealso{
shape_ebstrata
}
