% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erieshore.R
\name{base_erieshore}
\alias{base_erieshore}
\title{base_ebstrata}
\format{
a ggplot object with the strata as a geom_sf() object
}
\description{
Lake Erie shoreline as base ggplot object
}
\examples{
\dontrun{
data("base_erieshore")
base_erieshore
}

}
\seealso{
shape_erieshore
}
