% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jurisdictions.R
\name{base_jurisdictions}
\alias{base_jurisdictions}
\title{base_jurisdictions}
\format{
a ggplot object with the states and province as a geom_sf() object
}
\description{
Lake Erie jurisdictions as base ggplot object
}
\examples{
\dontrun{
data("base_jurisdictions")
base_jurisdictions
}

}
\seealso{
shape_erieshore
}
