% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lestrata.R
\name{base_lestrata}
\alias{base_lestrata}
\title{base_lestrata}
\format{
a ggplot object with the strata as a geom_sf() object
}
\description{
Lake Erie Strata as base ggplot object
}
\examples{
\dontrun{
data("base_lestrata")
base_lestrata
}

}
\seealso{
shape_lestrata
}
