% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbstrata.R
\name{base_wbstrata}
\alias{base_wbstrata}
\title{base_wbstrata}
\format{
a ggplot object with the strata as a geom_sf() object
}
\description{
West Basin Strata as base ggplot object
}
\examples{
\dontrun{
data("base_wbstrata")
base_wbstrata
}

}
\seealso{
shape_wbstrata
}
