% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5mincent_surv.R
\name{shape_5mincent_surv}
\alias{shape_5mincent_surv}
\title{shape_5mincent_surv}
\format{
an sf spatial data frame
}
\description{
Lake Erie 5-minute grid centroids within survey area as shapefile
}
\examples{
\dontrun{
data("shape_5mincent_surv")
class(shape_5mincent_surv)
plot(shape_5mincent_surv)
}
}
\seealso{
base_5mincent_surv
}
