% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5mingrid.R
\name{shape_5mingrid}
\alias{shape_5mingrid}
\title{shape_5mingrid}
\format{
an sf spatial data frame
}
\description{
Lake Erie 5-minute grid as shapefile
}
\examples{
\dontrun{
data("shape_5mingrid")
class(shape_5mingrid)
plot(shape_5mingrid)
}
}
\seealso{
base_5mingrid
}
