% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5mingrid_surv.R
\name{shape_5mingrid_surv}
\alias{shape_5mingrid_surv}
\title{shape_5mingrid_surv}
\format{
an sf spatial data frame
}
\description{
Lake Erie 5-minute grid within survey area as shapefile
}
\examples{
\dontrun{
data("shape_5mingrid_surv")
class(shape_5mingrid_surv)
plot(shape_5mingrid_surv)
}
}
\seealso{
base_5mingrid_surv
}
