% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbstrata.R
\name{shape_cbstrata}
\alias{shape_cbstrata}
\title{shape_cbstrata}
\format{
an sf spatial data frame
}
\description{
Central Basin Strata as shapefile
}
\examples{
\dontrun{
data("shape_cbstrata")
class(shape_cbstrata)
plot(shape_cbstrata)
}
}
\seealso{
base_cbstrata
}
