% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebstrata.R
\name{shape_ebstrata}
\alias{shape_ebstrata}
\title{shape_ebstrata}
\format{
an sf spatial data frame
}
\description{
East Basin Strata as shapefile
}
\examples{
\dontrun{
data("shape_ebstrata")
class(shape_ebstrata)
plot(shape_ebstrata)
}
}
\seealso{
base_ebstrata
}
