% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erieshore.R
\name{shape_erieshore}
\alias{shape_erieshore}
\title{shape_erieshore}
\format{
an sf spatial data frame
}
\description{
Lake Erie shoreline as shapefile
}
\examples{
\dontrun{
data("shape_erieshore")
class(shape_erieshore)
plot(shape_erieshore)
}
}
\seealso{
base_erieshore
}
