% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jurisdictions.R
\name{shape_jurisdictions}
\alias{shape_jurisdictions}
\title{shape_jurisdictions}
\format{
an sf spatial data frame
}
\description{
Lake Erie states and province as shapefile. Data from Natural Earth (public domain, available:naturalearthdata.com).
}
\examples{
\dontrun{
data("shape_jurisdictions")
class(shape_jurisdictions)
plot(shape_jurisdictions)
}
}
\seealso{
shape_jurisdictions
}
