% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lestrata.R
\name{shape_lestrata}
\alias{shape_lestrata}
\title{shape_lestrata}
\format{
an sf spatial data frame
}
\description{
Lake Erie Strata as shapefile
}
\examples{
\dontrun{
data("shape_lestrata")
class(shape_lestrata)
plot(shape_lestrata)
}
}
\seealso{
base_lestrata
}
