% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbstrata.R
\name{shape_wbstrata}
\alias{shape_wbstrata}
\title{shape_wbstrata}
\format{
an sf spatial data frame
}
\description{
West Basin Strata as shapefile
}
\examples{
\dontrun{
data("shape_wbstrata")
class(shape_wbstrata)
plot(shape_wbstrata)
}
}
\seealso{
base_wbstrata
}
